+++ source/Nuke.Core/Execution/TargetDefinitionLoader.cs
@@ -7,18 +7,34 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using Nuke.Core.Utilities;
+using Nuke.Core.Utilities.Collections;
 
 namespace Nuke.Core.Execution
 {
     internal static class TargetDefinitionLoader
     {
-        public static IReadOnlyCollection<TargetDefinition> GetExecutionList(NukeBuild build)
+        public static IReadOnlyCollection<TargetDefinition> GetExecutingTargets(NukeBuild build)
         {
             ControlFlow.Assert(build.TargetDefinitions.All(x => !x.Name.EqualsOrdinalIgnoreCase(BuildExecutor.DefaultTarget)),
                 "The name 'default' cannot be used as target name.");
 
-            var specifiedTargets = build.InvokedTargets.Select(x => GetDefinition(x, build)).ToList();
-            return GetSortedList(build, specifiedTargets);
+            var invokedTargets = build.InvokedTargets.Select(x => GetDefinition(x, build)).ToList();
+            var executingTargets = GetUnfilteredExecutingTargets(build, invokedTargets);
+            var skippedTargets = executingTargets
+                    .Where(x => !invokedTargets.Contains(x) &&
+                                build.SkippedTargets != null &&
+                                (build.SkippedTargets.Length == 0 ||
+                                 build.SkippedTargets.Contains(x.Name, StringComparer.OrdinalIgnoreCase))).ToList();
+            skippedTargets.ForEach(x => x.Conditions.Add(() => false));
+
+            string[] GetNames (IEnumerable<TargetDefinition> targets)
+                => targets.Select(x => x.Name).ToArray();
+            
+            ReflectionService.SetValue(build, nameof(NukeBuild.InvokedTargets), GetNames(invokedTargets));
+            ReflectionService.SetValue(build, nameof(NukeBuild.SkippedTargets), GetNames(skippedTargets));
+            ReflectionService.SetValue(build, nameof(NukeBuild.ExecutingTargets), GetNames(executingTargets.Except(skippedTargets)));
+
+            return executingTargets;
         }
 
         private static TargetDefinition GetDefinition(
@@ -42,14 +58,14 @@ namespace Nuke.Core.Execution
             return targetDefinition;
         }
 
-        private static List<TargetDefinition> GetSortedList(NukeBuild build, IReadOnlyCollection<TargetDefinition> specifiedTargets)
+        private static List<TargetDefinition> GetUnfilteredExecutingTargets(NukeBuild build, IReadOnlyCollection<TargetDefinition> invokedTargets)
         {
             var vertexDictionary = build.TargetDefinitions.ToDictionary(x => x, x => new Vertex<TargetDefinition>(x));
             foreach (var pair in vertexDictionary)
                 pair.Value.Dependencies = pair.Key.TargetDefinitionDependencies.Select(x => vertexDictionary[x]).ToList();
 
             var graphAsList = vertexDictionary.Values.ToList();
-            var result = new List<TargetDefinition>();
+            var executingTargets = new List<TargetDefinition>();
 
             while (graphAsList.Any())
             {
@@ -79,14 +95,17 @@ namespace Nuke.Core.Execution
                 graphAsList.Remove(independent);
 
                 var targetDefinition = independent.Value;
-                var executableDependency = result.SelectMany(x => x.TargetDefinitionDependencies).Contains(targetDefinition)
-                                           && !EnvironmentInfo.ArgumentSwitch("nodeps");
-                if (specifiedTargets.Contains(targetDefinition) || executableDependency)
-                    result.Add(targetDefinition);
+                var dependencies = executingTargets.SelectMany(x => x.TargetDefinitionDependencies);
+                if (!invokedTargets.Contains(targetDefinition) &&
+                    !dependencies.Contains(targetDefinition))
+                    continue;
+
+                executingTargets.Add(targetDefinition);
             }
 
-            result.Reverse();
-            return result;
+            executingTargets.Reverse();
+
+            return executingTargets;
         }
     }
 }